chore: refactor ValueLabel out of class
